pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leonardo-assistant"
version = "0.2.0"
description = "Voice assistant with client-server architecture (Raspberry Pi + Mini PC)"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
client = [
    "sounddevice>=0.5",
    "webrtcvad>=2.0.10",
    "openwakeword>=0.6",
]
server = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "faster-whisper>=1.0",
    "requests>=2.31",
    "piper-tts>=1.4",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
]

[project.scripts]
leonardo-client = "client.main:main"
leonardo-server = "server.main:main"

[tool.setuptools.packages.find]
include = ["shared*", "client*", "server*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
